+++ pkg/tracingpolicy/generictracingpolicy.go
@@ -11,22 +11,23 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"sigs.k8s.io/yaml"
 
+	"github.com/cilium/tetragon/pkg/crdutils"
 	"github.com/cilium/tetragon/pkg/k8s/apis/cilium.io/client"
 	"github.com/cilium/tetragon/pkg/k8s/apis/cilium.io/v1alpha1"
 )
 
 var (
-	TPContext  *CRDContext[*GenericTracingPolicy]
-	TPNContext *CRDContext[*GenericTracingPolicyNamespaced]
+	TPContext  *crdutils.CRDContext[*GenericTracingPolicy]
+	TPNContext *crdutils.CRDContext[*GenericTracingPolicyNamespaced]
 )
 
 func init() {
 	var err error
-	TPContext, err = NewCRDContext[*GenericTracingPolicy](&client.TracingPolicyCRD.Definition)
+	TPContext, err = crdutils.NewCRDContext[*GenericTracingPolicy](&client.TracingPolicyCRD.Definition)
 	if err != nil {
 		panic(fmt.Sprintf("failed to initialize CRD context for TracingPolicy: %v", err))
 	}
-	TPNContext, err = NewCRDContext[*GenericTracingPolicyNamespaced](&client.TracingPolicyNamespacedCRD.Definition)
+	TPNContext, err = crdutils.NewCRDContext[*GenericTracingPolicyNamespaced](&client.TracingPolicyNamespacedCRD.Definition)
 	if err != nil {
 		panic(fmt.Sprintf("failed to initialize CRD context for TracingPolicyNamespaced: %v", err))
 	}
